pinger/requirements.txt
altgraph==0.17.4
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-admin==2.0.2
django-cors-headers==4.7.0
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
excel-base==1.0.4
idna==3.10
isoweek==1.3.3
packaging==25.0
pillow==10.2.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.31.0
psycopg2-binary==2.9.9